// src/types/cats.ts
export type Cat = {
  id: string;
  image: string;
  name: string;
  description: string;
  weight: string;
  life_span: string;
  sheeding: number;
  child_friendly: number;
  stranger_friendly: number;
  dog_friendly: number;
  playfulness: number;
  intelligence: number;
  temperament: string;
  indoors: number;
  grooming: number;
  social_needs: number;
  activity_level: number;
  shedding_level: number;
  health_issues: number;
  vocalisation: number;
  hairless: number;
  rare: number;
  wikipedia_url: string | null;
  hypoallergenic: number;
  short_legs: number;
  suppresed_tail: number;
  affectionate: number;
};